refactor: extract trait impls and utilities to impls.rs and utilities.rs
Moves the remaining post-pallet-module code out of lib.rs:
- impls.rs: BrokerDeregistrationCheck, FlipBurnOrMoveInfo, SwapParameterValidation,
  and AffiliateRegistry trait implementations
- utilities.rs: hard_coded_price_for_asset and split_off_highest_impact_swap helpers
