+++ state-chain/pallets/cf-swapping/src/lib.rs
@@ -79,13 +79,16 @@ mod benchmarking;
 mod dca;
 mod execution;
 mod fees;
+mod impls;
 mod swap_state;
+pub mod utilities;
 
 pub mod migrations;
 pub mod weights;
 pub use weights::WeightInfo;
 pub use dca::DcaState;
 pub use fees::{BrokerFeesTracker, FeeRateAndMinimum, NetworkFeeTracker};
+pub use impls::BrokerDeregistrationCheck;
 pub use swap_state::{Stage1, Stage2, Stage3, Stage4, Stage5, StageFailed, SwapState};
 pub(crate) use swap_state::{GroupSwapState, SwapGroupPair};
 
@@ -1581,195 +1584,3 @@ pub mod pallet {
 	}
 }
 
-pub struct BrokerDeregistrationCheck<T>(PhantomData<T>);
-
-impl<T: Config> DeregistrationCheck for BrokerDeregistrationCheck<T> {
-	type AccountId = T::AccountId;
-	type Error = Error<T>;
-
-	fn check(account_id: &Self::AccountId) -> Result<(), Self::Error> {
-		ensure!(
-			!BrokerPrivateBtcChannels::<T>::contains_key(account_id),
-			Error::<T>::PrivateChannelExistsForBroker
-		);
-		ensure!(
-			AffiliateAccountDetails::<T>::iter_key_prefix(account_id).all(|affiliate_account_id| {
-				T::BalanceApi::free_balances(&affiliate_account_id)
-					.iter()
-					.all(|(_, amount)| *amount == 0)
-			}),
-			Error::<T>::AffiliateEarnedFeesNotWithdrawn
-		);
-
-		Ok(())
-	}
-}
-
-impl<T: Config> cf_traits::FlipBurnOrMoveInfo for Pallet<T> {
-	fn take_flip_to_burn() -> i128 {
-		FlipToBurn::<T>::take()
-	}
-	fn take_flip_to_be_sent_to_gateway() -> AssetAmount {
-		FlipToBeSentToGateway::<T>::take()
-	}
-}
-
-impl<T: Config> SwapParameterValidation for Pallet<T> {
-	type AccountId = T::AccountId;
-
-	fn get_swap_limits() -> cf_traits::SwapLimits {
-		cf_traits::SwapLimits {
-			max_swap_retry_duration_blocks: MaxSwapRetryDurationBlocks::<T>::get(),
-			max_swap_request_duration_blocks: MaxSwapRequestDurationBlocks::<T>::get(),
-		}
-	}
-
-	fn validate_refund_params(
-		input_asset: Asset,
-		output_asset: Asset,
-		retry_duration: BlockNumber,
-		max_oracle_price_slippage: Option<BasisPoints>,
-	) -> Result<(), DispatchError> {
-		// Check that the retry duration is within limits.
-		let max_swap_retry_duration_blocks = MaxSwapRetryDurationBlocks::<T>::get();
-		if retry_duration > max_swap_retry_duration_blocks {
-			return Err(DispatchError::from(Error::<T>::RetryDurationTooHigh));
-		}
-
-		// Check that the oracle prices are available for the assets.
-		if let Some(_max_oracle_price_slippage) = max_oracle_price_slippage {
-			if T::PriceFeedApi::get_price(input_asset).is_none() ||
-				T::PriceFeedApi::get_price(output_asset).is_none()
-			{
-				return Err(DispatchError::from(Error::<T>::OraclePriceNotAvailable));
-			}
-		}
-
-		Ok(())
-	}
-
-	fn validate_dca_params(params: &cf_primitives::DcaParameters) -> Result<(), DispatchError> {
-		let max_swap_request_duration_blocks = MaxSwapRequestDurationBlocks::<T>::get();
-
-		if params.number_of_chunks != 1 {
-			if params.number_of_chunks == 0 {
-				return Err(DispatchError::from(Error::<T>::ZeroNumberOfChunksNotAllowed));
-			}
-			if params.chunk_interval == 0 {
-				return Err(DispatchError::from(Error::<T>::ChunkIntervalTooLow));
-			}
-			if let Some(total_swap_request_duration) =
-				params.number_of_chunks.saturating_sub(1).checked_mul(params.chunk_interval)
-			{
-				if total_swap_request_duration > max_swap_request_duration_blocks {
-					return Err(DispatchError::from(Error::<T>::SwapRequestDurationTooLong));
-				}
-			} else {
-				return Err(DispatchError::from(Error::<T>::InvalidDcaParameters));
-			}
-		}
-		Ok(())
-	}
-
-	fn validate_broker_fees(
-		broker_fees: &Beneficiaries<Self::AccountId>,
-	) -> Result<(), DispatchError> {
-		let total_bps = broker_fees
-			.iter()
-			.fold(0, |total, Beneficiary { bps, .. }| total.saturating_add(*bps));
-
-		ensure!(total_bps <= 1000, Error::<T>::BrokerCommissionBpsTooHigh);
-
-		Ok(())
-	}
-
-	fn get_minimum_vault_swap_fee_for_broker(broker_id: &Self::AccountId) -> BasisPoints {
-		VaultSwapMinimumBrokerFee::<T>::get(broker_id)
-	}
-}
-
-impl<T: Config> AffiliateRegistry for Pallet<T> {
-	type AccountId = T::AccountId;
-
-	fn get_account_id(
-		broker_id: &Self::AccountId,
-		affiliate_short_id: AffiliateShortId,
-	) -> Option<Self::AccountId> {
-		AffiliateIdMapping::<T>::get(broker_id, affiliate_short_id)
-	}
-
-	/// This function iterates over a storage map. Only for use in rpc methods.
-	fn get_short_id(
-		broker_id: &Self::AccountId,
-		affiliate_id: &Self::AccountId,
-	) -> Option<AffiliateShortId> {
-		AffiliateAccountDetails::<T>::get(broker_id, affiliate_id).map(|details| details.short_id)
-	}
-
-	fn reverse_mapping(broker_id: &Self::AccountId) -> BTreeMap<Self::AccountId, AffiliateShortId> {
-		AffiliateIdMapping::<T>::iter_prefix(broker_id)
-			.map(|(short_id, account_id)| (account_id, short_id))
-			.collect()
-	}
-}
-
-pub mod utilities {
-	use super::*;
-
-	/// Provides a static price that can be used as a fallback when estimating fees/gas.
-	/// These prices are rough approximations of real market prices and should be updated
-	/// occasionally if the prices move significantly.
-	pub fn hard_coded_price_for_asset(asset: Asset) -> Price {
-		let price_usd_cents = match asset {
-			Asset::Usdc |
-			Asset::Usdt |
-			Asset::ArbUsdc |
-			Asset::ArbUsdt |
-			Asset::SolUsdc |
-			Asset::SolUsdt |
-			Asset::HubUsdc |
-			Asset::HubUsdt => 100, // $1
-			Asset::Flip => 40,                     // ~$0.40
-			Asset::Eth | Asset::ArbEth => 280_000, // ~$2,800
-			Asset::Dot | Asset::HubDot => 200,     // ~$2
-			Asset::Btc | Asset::Wbtc => 8_650_000, // ~$86,500
-			Asset::Sol => 12_700,                  // ~$127
-		};
-
-		Price::from_usd_cents(asset, price_usd_cents)
-	}
-
-	pub(super) fn split_off_highest_impact_swap<T: Config>(
-		swaps: &mut Vec<Swap<T>>,
-		failed_swap_group: Vec<SwapState<T, StageFailed>>,
-	) -> Option<Swap<T>> {
-		// Check invariants:
-		if failed_swap_group.is_empty() {
-			log_or_panic!(
-				"Invariant violation: there should be at least one swap in a failed group"
-			)
-		}
-		for failed_swap in &failed_swap_group {
-			if !swaps.iter().any(|swap| swap.swap_id == failed_swap.swap_id()) {
-				log_or_panic!(
-					"Invariant violation: failed group must be a subset of all executed swaps"
-				)
-			}
-		}
-		// Find a swap id that we want to remove (in theory there should always be
-		// one from the failing asset/direction, but if we don't for some reason, the fallback is to
-		// remove nothing, which would abort the entire batch):
-		let maybe_swap_id_to_remove = failed_swap_group
-			.iter()
-			// If the direction is TO_STABLE, swap amount is in the input amount of
-			// *the same* asset (swaps from different assets are executed separately).
-			// If the direction is FROM_STABLE, swap amount is the amount in USDC.
-			// Either way, the amounts are in the same asset, so we can compare them directly:
-			.max_by_key(|swap| swap.stage.swap_amount)
-			.map(|swap| swap.swap_id());
-
-		maybe_swap_id_to_remove.and_then(|swap_id_to_remove| {
-			swaps.extract_if(.., |swap| swap.swap_id == swap_id_to_remove).next()
-		})
-	}
-}
